% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_capCosts.R
\name{calculate_capCosts}
\alias{calculate_capCosts}
\title{Evaluate the costs of REMIND technology placeholders (e.g. te_eselt_pass_sm).}
\usage{
calculate_capCosts(base_price, Fdemand_ES, EDGE2CESmap, EDGE2teESmap,
  REMIND2ISO_MAPPING, REMINDyears, scenario)
}
\description{
Evaluate the costs of REMIND technology placeholders (e.g. te_eselt_pass_sm).
}
