% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRDS.R
\name{createRDS}
\alias{createRDS}
\title{Creates RDS files for iterative REMIND-EDGE runs from csv input files.}
\usage{
createRDS(input_path)
}
\description{
Creates RDS files for iterative REMIND-EDGE runs from csv input files.
}
