% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_prices.R
\name{merge_prices}
\alias{merge_prices}
\title{Merge REMIND-derived fuel prices with non-fuel costs.}
\usage{
merge_prices(gdx, REMINDmapping, REMINDyears, intensity_data,
  nonfuel_costs, module = "edge_esm")
}
\description{
Merge REMIND-derived fuel prices with non-fuel costs.
}
