% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare4REMIND.R
\name{prepare4REMIND}
\alias{prepare4REMIND}
\title{Prepare outputs for REMIND: Set correct tech names, CES node names and years.}
\usage{
prepare4REMIND(demByTech, intensity, capCost, EDGE2teESmap, REMINDtall)
}
\description{
Prepare outputs for REMIND: Set correct tech names, CES node names and years.
}
