% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shares_intensities_and_demand.R
\name{shares_intensity_and_demand}
\alias{shares_intensity_and_demand}
\title{Derive shares, demand and intensities from the logit tree.}
\usage{
shares_intensity_and_demand(logit_shares, MJ_km_base, REMIND2ISO_MAPPING,
  EDGE2CESmap, REMINDyears, scenario, demand_input = NULL)
}
\description{
Derive shares, demand and intensities from the logit tree.
}
