% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_logit.R
\name{calculate_logit}
\alias{calculate_logit}
\title{Merge REMIND-derived fuel prices with non-fuel costs.}
\usage{
calculate_logit(gdx, REMINDmapping, path2VOT, path2calibration_output,
  path2logitparam, path2intensities, path2UCD, full_data = F)
}
\description{
Merge REMIND-derived fuel prices with non-fuel costs.
}
