% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_logit.R
\name{calculate_logit}
\alias{calculate_logit}
\title{Merge REMIND-derived fuel prices with non-fuel costs.}
\usage{
calculate_logit(prices, vot_data, sw_data, logit_params, intensity_data,
  full_data = F)
}
\description{
Merge REMIND-derived fuel prices with non-fuel costs.
}
